Fixed Hiring business logic and Rest API validations

+++ app/Services/HireService.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace App\Services;
 
 use App\Http\Requests\DeveloperRequest;
@@ -7,8 +8,6 @@ use App\Models\Developer;
 use App\Models\Hire;
 use App\Services\Contracts\IHireService;
 use Exception;
-use http\Exception\UnexpectedValueException;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class HireService implements IHireService
@@ -20,7 +19,8 @@ class HireService implements IHireService
         return Hire::all();
     }
 
-    public function storeHire(HireRequest $request) {
+    public function storeHire(HireRequest $request)
+    {
         // Select all from developer where name = names from hire_developers
         $hire_devs_by_id = Developer::where('id', $request->ids)->get();
         $selected_developers_to_hire = $request->ids;
@@ -31,19 +31,20 @@ class HireService implements IHireService
             // For every selected developer: Select all from hire_developers where names = 'selected developer' where start_date and end_date do not overlap
             $select_hired_developers_by_id =
                 Hire::select('*')
-                    ->where('developer_id', '=', $single_developer)
-                    ->where(function ($query) use ($request) { $query->whereBetween('start_date', [$request->start_date, $request->end_date])->orWhereBetween('end_date', [$request->start_date, $request->end_date]); })
+                    ->where('developer_id', '=', $single_developer->id)
+                    ->where(function ($query) use ($request) {
+                        $query->whereBetween('start_date', [$request->start_date, $request->end_date])->orWhereBetween('end_date', [$request->start_date, $request->end_date]);
+                    })
                     ->first();
-//            $check_rows = $developer_for_hire_from_db_select;
             $store_hired_developers_by_names[] = $select_hired_developers_by_id;
 
-            if ($select_hired_developers_by_id) {
-                header('Location: /hire');
-                die("Select valid date");
-            }
             // Validation checks.
+            if ($select_hired_developers_by_id) {
+                throw new Exception("Developer '$single_developer->name' already hired for this time period.");
+            }
+
             if ($request->end_date < $request->start_date || $request->start_date < $now || $request->end_date < $now) {
-               throw new Exception("Select valid date");
+                throw new Exception("Select valid date");
             }
 
             // Check if the looped developers count is equal to the selected developers from the list: if more than one developer is selected from the list, the loop does not exit but gets back and iterates for every other.
@@ -52,23 +53,29 @@ class HireService implements IHireService
             }
             // Loop all the collected/selected in array developers and hire simultaneously.
             foreach ($collected_developers as $single_collected_developer) {
-                foreach($hire_devs_by_id as $dev) {
+                foreach ($hire_devs_by_id as $dev) {
                     Hire::insert(
-                        ["names" => $dev->name, "developer_id" => $dev->id, "start_date" => $request->start_date, "end_date" => $request->end_date, 'user_hired_id'=> Auth::user()->id]
+                        ["names" => $dev->name, "developer_id" => $dev->id, "start_date" => $request->start_date, "end_date" => $request->end_date, 'user_hired_id' => Auth::user()->id]
                     );
                 }
             }
         }
     }
 
-    public function deleteHire($id) {
-        $delete_dev = Hire::findOrFail($id);
-        $delete_dev->delete();
+    public function deleteHire($id)
+    {
+        $delete_dev = Hire::find($id);
+        if (is_null($delete_dev)) {
+            return false;
+        }
+
+        return $delete_dev->delete();
     }
 
-    public function updateHires(DeveloperRequest $request, $id) {
+    public function updateHires(DeveloperRequest $request, $id)
+    {
         $hire = Hire::where('developer_id', $id)->get();
-        foreach($hire as $single_hire) {
+        foreach ($hire as $single_hire) {
             $single_hire->names = $request->get('name', 'No Data');
             $single_hire->save();
         }
